Move extended resource test to a new file
