+++ test/suites/integration/scheduling_test.go
@@ -20,12 +20,9 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/samber/lo"
-	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
 	"knative.dev/pkg/ptr"
 
 	"github.com/aws/karpenter-core/pkg/apis/provisioning/v1alpha5"
@@ -208,220 +205,4 @@ var _ = Describe("Scheduling", func() {
 		Expect(ptr.StringValue(env.GetInstance(pod.Spec.NodeName).InstanceType)).To(Equal("c4.large"))
 		Expect(env.GetNode(pod.Spec.NodeName).Labels[v1alpha5.ProvisionerNameLabelKey]).To(Equal(provisionerHighPri.Name))
 	})
-	Context("Extended Resources", func() {
-		It("should provision nodes for a deployment that requests nvidia.com/gpu", func() {
-			ExpectNvidiaDevicePluginCreated()
-			DeferCleanup(func() {
-				ExpectNvidiaDevicePluginDeleted()
-			})
-
-			provider := awstest.AWSNodeTemplate(v1alpha1.AWSNodeTemplateSpec{AWS: v1alpha1.AWS{
-				SecurityGroupSelector: map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-				SubnetSelector:        map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-			}})
-			provisioner := test.Provisioner(test.ProvisionerOptions{
-				ProviderRef: &v1alpha5.ProviderRef{Name: provider.Name},
-				Requirements: []v1.NodeSelectorRequirement{
-					{
-						Key:      v1alpha1.LabelInstanceCategory,
-						Operator: v1.NodeSelectorOpExists,
-					},
-				},
-			})
-			numPods := 1
-			dep := test.Deployment(test.DeploymentOptions{
-				Replicas: int32(numPods),
-				PodOptions: test.PodOptions{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{"app": "large-app"},
-					},
-					ResourceRequirements: v1.ResourceRequirements{
-						Requests: v1.ResourceList{
-							"nvidia.com/gpu": resource.MustParse("1"),
-						},
-						Limits: v1.ResourceList{
-							"nvidia.com/gpu": resource.MustParse("1"),
-						},
-					},
-				},
-			})
-			selector := labels.SelectorFromSet(dep.Spec.Selector.MatchLabels)
-			env.ExpectCreated(provisioner, provider, dep)
-			env.EventuallyExpectHealthyPodCount(selector, numPods)
-			env.ExpectCreatedNodeCount("==", 1)
-			env.EventuallyExpectCreatedNodesInitialized()
-		})
-		It("should provision nodes for a deployment that requests nvidia.com/gpu (Bottlerocket)", func() {
-			// For Bottlerocket, we are testing that resources are initialized without needing a device plugin
-			provider := awstest.AWSNodeTemplate(v1alpha1.AWSNodeTemplateSpec{AWS: v1alpha1.AWS{
-				AMIFamily:             &v1alpha1.AMIFamilyBottlerocket,
-				SecurityGroupSelector: map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-				SubnetSelector:        map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-			}})
-			provisioner := test.Provisioner(test.ProvisionerOptions{
-				ProviderRef: &v1alpha5.ProviderRef{Name: provider.Name},
-				Requirements: []v1.NodeSelectorRequirement{
-					{
-						Key:      v1alpha1.LabelInstanceCategory,
-						Operator: v1.NodeSelectorOpExists,
-					},
-				},
-			})
-			numPods := 1
-			dep := test.Deployment(test.DeploymentOptions{
-				Replicas: int32(numPods),
-				PodOptions: test.PodOptions{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{"app": "large-app"},
-					},
-					ResourceRequirements: v1.ResourceRequirements{
-						Requests: v1.ResourceList{
-							"nvidia.com/gpu": resource.MustParse("1"),
-						},
-						Limits: v1.ResourceList{
-							"nvidia.com/gpu": resource.MustParse("1"),
-						},
-					},
-				},
-			})
-			selector := labels.SelectorFromSet(dep.Spec.Selector.MatchLabels)
-			env.ExpectCreated(provisioner, provider, dep)
-			env.EventuallyExpectHealthyPodCount(selector, numPods)
-			env.ExpectCreatedNodeCount("==", 1)
-			env.EventuallyExpectCreatedNodesInitialized()
-		})
-		It("should provision nodes for a deployment that requests vpc.amazonaws.com/pod-eni (security groups for pods)", func() {
-			ExpectPodENIEnabled()
-			DeferCleanup(func() {
-				ExpectPodENIDisabled()
-			})
-			env.ExpectSettingsOverridden(map[string]string{
-				"aws.enablePodENI": "true",
-			})
-			provider := awstest.AWSNodeTemplate(v1alpha1.AWSNodeTemplateSpec{AWS: v1alpha1.AWS{
-				SecurityGroupSelector: map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-				SubnetSelector:        map[string]string{"karpenter.sh/discovery": settings.FromContext(env.Context).ClusterName},
-			}})
-			provisioner := test.Provisioner(test.ProvisionerOptions{
-				ProviderRef: &v1alpha5.ProviderRef{Name: provider.Name},
-				Requirements: []v1.NodeSelectorRequirement{
-					{
-						Key:      v1alpha1.LabelInstanceCategory,
-						Operator: v1.NodeSelectorOpExists,
-					},
-				},
-			})
-			numPods := 1
-			dep := test.Deployment(test.DeploymentOptions{
-				Replicas: int32(numPods),
-				PodOptions: test.PodOptions{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{"app": "large-app"},
-					},
-					ResourceRequirements: v1.ResourceRequirements{
-						Requests: v1.ResourceList{
-							"vpc.amazonaws.com/pod-eni": resource.MustParse("1"),
-						},
-						Limits: v1.ResourceList{
-							"vpc.amazonaws.com/pod-eni": resource.MustParse("1"),
-						},
-					},
-				},
-			})
-			selector := labels.SelectorFromSet(dep.Spec.Selector.MatchLabels)
-			env.ExpectCreated(provisioner, provider, dep)
-			env.EventuallyExpectHealthyPodCount(selector, numPods)
-			env.ExpectCreatedNodeCount("==", 1)
-			env.EventuallyExpectCreatedNodesInitialized()
-		})
-	})
 })
-
-func ExpectNvidiaDevicePluginCreated() {
-	env.ExpectCreatedWithOffset(1, &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "nvidia-device-plugin-daemonset",
-			Namespace: "kube-system",
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{
-					"name": "nvidia-device-plugin-ds",
-				},
-			},
-			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
-				Type: appsv1.RollingUpdateDaemonSetStrategyType,
-			},
-			Template: v1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						"name": "nvidia-device-plugin-ds",
-					},
-				},
-				Spec: v1.PodSpec{
-					Tolerations: []v1.Toleration{
-						{
-							Key:      "nvidia.com/gpu",
-							Operator: v1.TolerationOpExists,
-							Effect:   v1.TaintEffectNoSchedule,
-						},
-					},
-					PriorityClassName: "system-node-critical",
-					Containers: []v1.Container{
-						{
-							Name:  "nvidia-device-plugin-ctr",
-							Image: "nvcr.io/nvidia/k8s-device-plugin:v0.12.3",
-							Env: []v1.EnvVar{
-								{
-									Name:  "FAIL_ON_INIT_ERROR",
-									Value: "false",
-								},
-							},
-							SecurityContext: &v1.SecurityContext{
-								AllowPrivilegeEscalation: lo.ToPtr(false),
-								Capabilities: &v1.Capabilities{
-									Drop: []v1.Capability{"ALL"},
-								},
-							},
-							VolumeMounts: []v1.VolumeMount{
-								{
-									Name:      "device-plugin",
-									MountPath: "/var/lib/kubelet/device-plugins",
-								},
-							},
-						},
-					},
-					Volumes: []v1.Volume{
-						{
-							Name: "device-plugin",
-							VolumeSource: v1.VolumeSource{
-								HostPath: &v1.HostPathVolumeSource{
-									Path: "/var/lib/kubelet/device-plugins",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	})
-}
-
-func ExpectNvidiaDevicePluginDeleted() {
-	env.ExpectDeletedWithOffset(1, &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "nvidia-device-plugin-daemonset",
-			Namespace: "kube-system",
-		},
-	})
-}
-
-func ExpectPodENIEnabled() {
-	env.ExpectDaemonSetEnvironmentVariableUpdatedWithOffset(1, types.NamespacedName{Namespace: "kube-system", Name: "aws-node"},
-		"ENABLE_POD_ENI", "true")
-}
-
-func ExpectPodENIDisabled() {
-	env.ExpectDaemonSetEnvironmentVariableUpdatedWithOffset(1, types.NamespacedName{Namespace: "kube-system", Name: "aws-node"},
-		"ENABLE_POD_ENI", "false")
-}
